% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{fug_bw}
\alias{fug_bw}
\title{Black and white filter}
\usage{
fug_bw(p, threshold = "auto")
}
\arguments{
\item{threshold}{Set a threshold for black vs. white. Defaults to "auto". See \link{threshold}.}

\item{plot}{An object or expression that evaluates as a plot.}
}
\description{
Run a grayscale filter on a plot
}

