% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{fug_fun}
\alias{fug_fun}
\title{Run an arbitrary function from imager}
\usage{
fug_fun(p, imager_fun, ...)
}
\arguments{
\item{p}{Expression that evaluates to a plot}

\item{imager_fun}{imager function}

\item{...}{Arguments to pass to \code{imager_fun}}
}
\description{
Run an arbitrary function from imager
}

