% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize.R
\name{rasterize}
\alias{rasterize}
\title{Creates a static image of each plot to pass to imager}
\usage{
rasterize(p)
}
\arguments{
\item{p}{Expression that evaluates to a plot}
}
\description{
Creates a static image of each plot to pass to imager
}

